simplify testing apparatus, split handleFieldValueChange function to its own file
