+++ app/domains/kontopfaendung/pkonto/antrag/pages.ts
@@ -1,9 +1,5 @@
-import { bankNameFromIBAN } from "~/components/formElements/inputs/iban/bankNameFromIBAN";
-import { fetchBanks } from "~/components/formElements/inputs/iban/fetchBanks";
-import type {
-  FieldValueChangeHandlerProps,
-  PagesConfig,
-} from "~/domains/pageSchemas";
+import { setBankNameFromIban } from "~/domains/kontopfaendung/services/setBankNameFromIban";
+import type { PagesConfig } from "~/domains/pageSchemas";
 import { translations } from "~/services/translations/translations";
 import { checkedRequired } from "~/services/validation/checkedCheckbox";
 import { emailSchema } from "~/services/validation/email";
@@ -14,8 +10,6 @@ import { schemaOrEmptyString } from "~/services/validation/schemaOrEmptyString";
 import { stringOptionalSchema } from "~/services/validation/stringOptional";
 import { stringRequiredSchema } from "~/services/validation/stringRequired";
 import { YesNoAnswer } from "~/services/validation/YesNoAnswer";
-
-const banks = await fetchBanks();
 
 export const kontopfaendungPkontoAntragPages = {
   grundvoraussetzungenDatenverarbeitung: {
@@ -44,34 +38,7 @@ export const kontopfaendungPkontoAntragPages = {
     },
     controlledFieldConfig: {
       fieldName: "bankName",
-      handleFieldValueChange: async ({
-        originalValue,
-        value,
-        controlledField,
-        setControlledFieldSrValue,
-      }: FieldValueChangeHandlerProps) => {
-        // needed to ensure value isn't automatically set upon initial render
-        if (originalValue !== value) {
-          if (value && typeof value === "string" && value.length > 0 && banks) {
-            // Debounce needed to not clobber the screen reader while typing
-            const timeout = setTimeout(() => {
-              const matchedBankName = bankNameFromIBAN(value, banks);
-              if (matchedBankName) {
-                setControlledFieldSrValue(matchedBankName);
-                controlledField.setValue(matchedBankName);
-                controlledField.validate();
-              } else {
-                setControlledFieldSrValue("");
-                controlledField.setValue("");
-              }
-            }, 1000);
-
-            return () => clearTimeout(timeout);
-          }
-          setControlledFieldSrValue("");
-          controlledField?.setValue("");
-        }
-      },
+      handleFieldValueChange: setBankNameFromIban,
       getScreenReaderAnnouncementText: (controlledFieldSrValue: string) =>
         `${translations.iban.bankIdentified.de}: ${controlledFieldSrValue}`,
     },
